ADFA-1375 | Refactor: Move EventBus registration to EmptyStateFragment
- Moved EventBus registration and unregistration logic from `LogViewFragment` to the base class `EmptyStateFragment`.
- The `onEditorLongPressed` method in `EmptyStateFragment` now calls `onFragmentLongPressed`, which is an abstract method implemented by subclasses like `BuildOutputFragment`.
- This change centralizes the long-press event handling for fragments that use an editor and have an empty state.
