+++ app/src/main/java/com/itsaky/androidide/fragments/debug/DebuggerFragment.kt
@@ -156,6 +156,9 @@ class DebuggerFragment :
         binding.pager.adapter = DebuggerPagerAdapter(this, tabs.map { it.second })
         mediator.attach()
     }
+    override fun onFragmentLongPressed() {
+        //TODO be defined
+    }
 }
 
 class DebuggerPagerAdapter(
