+++ idetooltips/src/main/java/com/itsaky/androidide/idetooltips/TooltipTag.kt
@@ -24,5 +24,6 @@ object TooltipTag {
     const val PROJECT_DIAGNOSTICS = "project.diagnostics"
     const val OUTPUT_CLEAR = "output.clear"
     const val OUTPUT_SHARE_EXTERNAL = "output.share.external"
+    const val PROJECT_BUILD_OUTPUT = "project.buildoutput"
 
 }
